requirements.txt
Flask==3.0.3
flask_cors
requests
python-dotenv
os